Initial commit.
Starting my journey into "Types and Programming Languages". It's
definitely a change of taste from the more logic-oriented, more
theoretical books "Proofs and Types" of Girard and "Programming in
Martin-Löf Type Theory".

I decided it would be worthfile to document this journey by writing down
the exercises in LaTeX (for as many as time and energy allow), and to
formalize the toy languages described in the book in Idris.