Missing eval rules + rule fixed.
I completely forgot to add the evaluation rules for the booleans to
`EvalsTo` last time and also made a small typo in one rule: these are
now fixed.

I've also removed the scaffolding code for my first proof attempt as the
proof seems to be way more difficult than I thought and the strategy
probably needs to be different.