Some headway with solution of 3.5.17.
The main idea of the (yet-to-be-done!) proof of 3.5.17, more precisely
of the implication

$$t \rightarrow^\ast v \Rightarrow t \downarrow v$$

(which is also the hard direction btw), is to recognize a derivation in
$\rightarrow^\ast$ as a value of the function `from_BInE_to_E`, i.e. to
*parse it* in terms of the rules formalized in the type `BInE`.
Once this is achieved, i.e. once the function `from_E_to_BInE` is
implemented, once can then implement `starImpliesBig` easily using this
function and induction, as the rules in `BInE` exactly mirror the rules
in `BigEvalsTo`.

So far, I've only been able to implement partially the function
`from_BInE_to_E` (which I thought should be *trivial*). This already
necessitated introducing some helper functions (notably, `(++)` and `map`).
Luckily, given these helper functions the implementation of
`from_BInE_to_E` wasn't that bad. Still, this exercise in formal proofs
(which I am very glad to have undertaken) indicates to me that formal
proofs are hard (or at least, seem to be hard given my level of
experience and understanding).