Progress with `from_E_to_BInE`.
The implementation of `from_E_to_BInE` is almost complete now.

In the course of doing so, I learned something very elementary but still
very valuable. Namely, I've always struggled with telling Idris that
a case is impossible given a value of type `Void`. My solution was
always to introduce this void value on the left hand side using
a `with`-clause. However, that is not necessary.

First of all, you can always use a simple `case`-clause for this

    case x of
         val impossible

But, even this is to complicated! Namely, given a value of type `Void`
you can simply use the function `absurd` to produce the necessary value.