Solution of 3.5.14 and some refactoring.
I've given a formal proof of exercises 3.5.14. This turned out to be way
more straight-forward than I feared[^1]. The only minor complication was
that because some of the constructors for `EvalsTo` took proofs as
arguments, I had to proof *equality of proofs* in certain cases in order
to show equality between one-step evaluation terms `r,r' : EvalsTo
t t'`.

[^1]: This fear being based upon the amount of time it took me to
formalize a solution of 3.5.17.