Added implicit variables.
As it turned out, the "weird" Idris errors I got were not (just)
a consequence of a persistent state in the Idris REPL, that is, I've
experienced this problem again. Moreover this time, I could not make it
go away by reloading the source files.

I fixed this problem now by adding some explicit type annotations for
implicit variables (all of type `Term`). I don't know why Idris did not
complain about this when I had the content of `Exercise_3_5_17.idr` in
the same file as `Arith.idr`, or why Idris was not able to infer the
type of these variables in the first place, but in any case, the problem
is now fixed.