Fully evaluated <=> Normal.
I've started proving that the two a priori different notions of being
`FullyEvaluated` and being `Normal` are in fact equivalent.

In the course of doing so, I also noted that I should have probably
defined and used (!) `Normal` already in `Ch03/Arith.idr`, instead of
using the ad hoc type that I used in `valuesAreNormal`.