"Fully evaluted => Normal" proven.
I've managed to show one direction of the equivalence between the two
notions. On the reverse direction I haven't been able to make much
progress, but I think that the prove should go roughly as follows:

* Show that a subterm of a normal term is still normal (either ad hoc,
or by introducing a subterm type).
* Recursively apply `normal_is_fully_evaluted` on the relevant subterm
and case-split on the result, treating the case of being a value and
being stuck separately.