Switched to explicit arguments.
I've changed the definition of the `Normal` type family to use
*explicit* arguments now (instead of *implicit* ones). This was
necessary because Idris' implementation of implicit arguments seems to
be broken, which is very unfortunate, and will probably not fixed
anytime soon (if ever). The two possible options at the moment seems to
be switching to Blodwen or to replace implicit arguments by explicit
ones. I've chosen the latter option for the moment, even though it makes
the code more verbose.