Inductive definitions of functions.
Since the argument is completely general, I've tried adding an
induction/fixed point principle, intended to be used to define the
`smallStep_eval` function.

To my surprise, I've encountered considerable difficulties in trying to
do so. I'm not sure what of that is due to bugs in Idris and what is due
   to my incomplete understanding.